+++ crates/iceberg/src/encryption/mod.rs
@@ -21,8 +21,8 @@
 //! for encrypting and decrypting data in Iceberg tables.
 
 mod crypto;
-pub(crate) mod encrypted_io;
-mod encryption_manager;
+pub(crate) mod io;
+mod manager;
 mod file_decryptor;
 mod file_encryptor;
 pub(crate) mod key_metadata;
@@ -30,8 +30,8 @@ pub mod kms;
 mod stream;
 
 pub use crypto::{AesGcmCipher, AesKeySize, SecureKey, SensitiveBytes};
-pub use encrypted_io::{EncryptedInputFile, EncryptedOutputFile};
-pub use encryption_manager::EncryptionManager;
+pub use io::{EncryptedInputFile, EncryptedOutputFile};
+pub use manager::EncryptionManager;
 pub use file_decryptor::AesGcmFileDecryptor;
 pub use file_encryptor::AesGcmFileEncryptor;
 pub use key_metadata::StandardKeyMetadata;
